Add files via upload
Combined the multiple source section of the current student quiz with the future students quiz, and moved the new, longer quiz to the profile section. Added weights to the "Which lawyer are you" quiz, and it now links to a seperate - currently blank - page depending on which type of lawyer scored the highest. Also resolved the warning messages caused by the storyboard elements not being aligned with their constraints.