Add files via upload
Created a data container singleton class. It holds the answers to the various quiz questions, and is each answer is globally accessible. It defaults to empty strings, and the data is written to memory, meaning that the information persists through closing the app.
Additionally, added a very WIP version of the About Me page that is will be accessible through the profile page. The intent is to have a scrolling page containing all the answers to the quiz questions, but it is very unfinished, and as such is not accessible from within the app itself